Moved all files into appropriate places. Added a babel plugin for root level imports
